Added upstream methods so view classes can be called from the StudentUI
